Tweak sorting and CSS
Take, for the purpose of sorting the resource list, the date of a
resource with several versions to be the date of the latest version.

Clip the EGAD Logo SVG to avoid doing it in CSS; it should hopefully
display better on a mobile phone now. A further spacing tweak on a
mobile.